+++ .bud/Build.cs
@@ -1,6 +1,4 @@
-using System.IO;
 using Bud;
-using Bud.Build;
 using Bud.CSharp;
 using Bud.Projects;
 using Bud.Resources;